Move logging logic from LogWriter to LogHelper
